requirements.txt
cffi==1.17.1
cryptography==44.0.2
pycparser==2.22
python-dotenv==1.1.0
